Removing styles to an exclusive file.
